+++ packages/vue/example/App.vue
@@ -2,7 +2,7 @@
 import ASCII from "./ASCII.vue"
 import LoginForm from "./LoginForm.vue"
 import Counter from "./Counter.vue"
-import StyledText from "./dist/Styled-Text.vue"
+import StyledText from "./Styled-Text.vue"
 import { ref } from "vue"
 import { getKeyHandler } from "@opentui/core"
